FOI letter request
Letter request for SpEd dataset  to DepEd.

The letter clearly states the intention to use the data in OpenStreetMap, in addition to asking for a formal letter of permission.

The request was approved, and the (revised) dataset was released on 3rd February. 

An initial release (excluding contact information) was made on 17th January, claiming they cannot release the contact details, under the Data Privacy Act. I responded, saying the Data Privacy Act is not applicable, since the request was asking for the institutional contact information, and not of any specific person.